refactor: replace inline revoke access functionality with a dedicated dialog component
- Removed inline alert dialog from ArtistAccessSelector and replaced it with the new RevokeAccessDialog component for better code organization and reusability.
- The RevokeAccessDialog handles the confirmation process for revoking file access, improving the user experience and maintaining separation of concerns.
